Find the relationship between the variables in a data set and fit that model

Finds the hirarchical relationship between all variables and connects them in a network such as to optimize the mutual information. This just fits the network selected by HillClimbing using a Bayesian information criteria (BIC).

discover_hierarchy_and_fit(data)

Arguments

data
the data.frame containing the observations and variables to inspect

Value

a character matrix with columns from and to representing the variables that are related

Examples

data(iris)
data(mtcars)
library(dplyr)
discover_hierarchy_and_fit(mtcars)

#> 
#>   Bayesian network parameters
#> 
#>   Parameters of node mpg (Gaussian distribution)
#> 
#> Conditional density: mpg | cyl + wt
#> Coefficients:
#> (Intercept)          cyl           wt  
#>   39.686261    -1.507795    -3.190972  
#> Standard deviation of the residuals: 2.567516 
#> 
#>   Parameters of node cyl (Gaussian distribution)
#> 
#> Conditional density: cyl
#> Coefficients:
#> (Intercept)  
#>      6.1875  
#> Standard deviation of the residuals: 1.785922 
#> 
#>   Parameters of node disp (Gaussian distribution)
#> 
#> Conditional density: disp | cyl
#> Coefficients:
#> (Intercept)          cyl  
#>  -156.60898     62.59893  
#> Standard deviation of the residuals: 54.38465 
#> 
#>   Parameters of node hp (Gaussian distribution)
#> 
#> Conditional density: hp | disp + wt + carb
#> Coefficients:
#> (Intercept)         disp           wt         carb  
#>  53.1671484    0.5166322  -28.5921378   23.5769117  
#> Standard deviation of the residuals: 24.32226 
#> 
#>   Parameters of node drat (Gaussian distribution)
#> 
#> Conditional density: drat | cyl + am
#> Coefficients:
#> (Intercept)          cyl           am  
#>   4.2233574   -0.1348772    0.5114038  
#> Standard deviation of the residuals: 0.3244723 
#> 
#>   Parameters of node wt (Gaussian distribution)
#> 
#> Conditional density: wt | disp
#> Coefficients:
#> (Intercept)         disp  
#> 1.599814597  0.007010325  
#> Standard deviation of the residuals: 0.4574133 
#> 
#>   Parameters of node qsec (Gaussian distribution)
#> 
#> Conditional density: qsec | hp + wt + vs
#> Coefficients:
#> (Intercept)           hp           wt           vs  
#> 16.02638440  -0.01767224   1.08994453   2.07551571  
#> Standard deviation of the residuals: 0.8156723 
#> 
#>   Parameters of node vs (Gaussian distribution)
#> 
#> Conditional density: vs | cyl + am
#> Coefficients:
#> (Intercept)          cyl           am  
#>   2.3181613   -0.2806444   -0.3548896  
#> Standard deviation of the residuals: 0.2620353 
#> 
#>   Parameters of node am (Gaussian distribution)
#> 
#> Conditional density: am | wt
#> Coefficients:
#> (Intercept)           wt  
#>   1.5424435   -0.3531567  
#> Standard deviation of the residuals: 0.3659338 
#> 
#>   Parameters of node gear (Gaussian distribution)
#> 
#> Conditional density: gear | am
#> Coefficients:
#> (Intercept)           am  
#>    3.210526     1.174089  
#> Standard deviation of the residuals: 0.4558808 
#> 
#>   Parameters of node carb (Gaussian distribution)
#> 
#> Conditional density: carb | mpg + cyl + gear
#> Coefficients:
#> (Intercept)          mpg          cyl         gear  
#>  -2.7816679   -0.1439035    0.3959199    1.6367526  
#> Standard deviation of the residuals: 0.8807524 
#> 

discover_hierarchy_and_fit(iris)

#> 
#>   Bayesian network parameters
#> 
#>   Parameters of node Sepal.Length (Gaussian distribution)
#> 
#> Conditional density: Sepal.Length | Sepal.Width + Petal.Length + Petal.Width
#> Coefficients:
#>  (Intercept)   Sepal.Width  Petal.Length   Petal.Width  
#>    1.8559975     0.6508372     0.7091320    -0.5564827  
#> Standard deviation of the residuals: 0.3145491 
#> 
#>   Parameters of node Sepal.Width (conditional Gaussian distribution)
#> 
#> Conditional density: Sepal.Width | Petal.Width + Species
#> Coefficients:
#>                      0          1          2
#> (Intercept)  3.2220507  1.3728630  1.6947730
#> Petal.Width  0.8371922  1.0536478  0.6314052
#> Standard deviation of the residuals:
#>         0          1          2  
#> 0.3724741  0.2370691  0.2747206  
#> Discrete parents' configurations:
#>       Species
#> 0      setosa
#> 1  versicolor
#> 2   virginica
#> 
#>   Parameters of node Petal.Length (conditional Gaussian distribution)
#> 
#> Conditional density: Petal.Length | Species
#> Coefficients:
#>                  0      1      2
#> (Intercept)  1.462  4.260  5.552
#> Standard deviation of the residuals:
#>         0          1          2  
#> 0.1736640  0.4699110  0.5518947  
#> Discrete parents' configurations:
#>       Species
#> 0      setosa
#> 1  versicolor
#> 2   virginica
#> 
#>   Parameters of node Petal.Width (conditional Gaussian distribution)
#> 
#> Conditional density: Petal.Width | Petal.Length + Species
#> Coefficients:
#>                         0            1            2
#> (Intercept)   -0.04822033  -0.08428835   1.13603130
#> Petal.Length   0.20124509   0.33105360   0.16029696
#> Standard deviation of the residuals:
#>         0          1          2  
#> 0.1004521  0.1233530  0.2627065  
#> Discrete parents' configurations:
#>       Species
#> 0      setosa
#> 1  versicolor
#> 2   virginica
#> 
#>   Parameters of node Species (multinomial distribution)
#> 
#> Conditional probability table:
#>      setosa versicolor  virginica 
#>  0.3333333  0.3333333  0.3333333 
#>