Encode a vector into a one hot data.frame

This function takes a vector of items and onehot encodes them into a data.frame which in reality is a tibble. Every unique value of the vector get converted into a column and the value is 1 for that column if the row number matches the position in the original vector. See the examples for details.

oneHotEncoder(x, f = function(y) y)

Arguments

x
the vector to convert to a one hot

f
a function to apply to x before encoding which defaults to identity

Value

the one hot encoded data.frame with the original data given in a Data column

Examples

library(lubridate)

#> 
#> Attaching package: ‘lubridate’

#> The following object is masked from ‘package:graph’:
#> 
#>     union

#> The following objects are masked from ‘package:BiocGenerics’:
#> 
#>     intersect, setdiff, union

#> The following objects are masked from ‘package:base’:
#> 
#>     date, intersect, setdiff, union

oneHotEncoder(
  x = seq(as.Date("2017-01-01"), by = "day", length.out = 10),
  f = function(x) wday(x, label = TRUE)
)

#>          Data Sun Mon Tue Wed Thu Fri Sat
#> 1  2017-01-01   1   0   0   0   0   0   0
#> 2  2017-01-02   0   1   0   0   0   0   0
#> 3  2017-01-03   0   0   1   0   0   0   0
#> 4  2017-01-04   0   0   0   1   0   0   0
#> 5  2017-01-05   0   0   0   0   1   0   0
#> 6  2017-01-06   0   0   0   0   0   1   0
#> 7  2017-01-07   0   0   0   0   0   0   1
#> 8  2017-01-08   1   0   0   0   0   0   0
#> 9  2017-01-09   0   1   0   0   0   0   0
#> 10 2017-01-10   0   0   1   0   0   0   0

oneHotEncoder(x = wday(seq(as.Date("2017-10-07"), by = "days", length.out = 10), label = TRUE))

#>    Data Sun Mon Tue Wed Thu Fri Sat
#> 1   Sat   0   0   0   0   0   0   1
#> 2   Sun   1   0   0   0   0   0   0
#> 3   Mon   0   1   0   0   0   0   0
#> 4   Tue   0   0   1   0   0   0   0
#> 5   Wed   0   0   0   1   0   0   0
#> 6   Thu   0   0   0   0   1   0   0
#> 7   Fri   0   0   0   0   0   1   0
#> 8   Sat   0   0   0   0   0   0   1
#> 9   Sun   1   0   0   0   0   0   0
#> 10  Mon   0   1   0   0   0   0   0