Strip a fitted GLM model

Reduce the size of a fitted GLM model by a massive factor. This is crucial in deploying R in production settings as the standard GLM is way to liberal with storing stuff we don't need for prediction.

stripGlmModel(a)

Arguments

a
the glm model object to strip

Value

a stripped GLM model

Examples

mylm <- glm(dist ~ speed, data = cars)
print(object.size(mylm))

#> 77792 bytes

mylm <- stripGlmModel(mylm)
print(object.size(mylm))

#> 49912 bytes

print(predict(mylm, cars))

#>         1         2         3         4         5         6         7         8 
#> -1.849460 -1.849460  9.947766  9.947766 13.880175 17.812584 21.744993 21.744993 
#>         9        10        11        12        13        14        15        16 
#> 21.744993 25.677401 25.677401 29.609810 29.609810 29.609810 29.609810 33.542219 
#>        17        18        19        20        21        22        23        24 
#> 33.542219 33.542219 33.542219 37.474628 37.474628 37.474628 37.474628 41.407036 
#>        25        26        27        28        29        30        31        32 
#> 41.407036 41.407036 45.339445 45.339445 49.271854 49.271854 49.271854 53.204263 
#>        33        34        35        36        37        38        39        40 
#> 53.204263 53.204263 53.204263 57.136672 57.136672 57.136672 61.069080 61.069080 
#>        41        42        43        44        45        46        47        48 
#> 61.069080 61.069080 61.069080 68.933898 72.866307 76.798715 76.798715 76.798715 
#>        49        50 
#> 76.798715 80.731124