Save a single CSV-table into a single table sqlite database

This function is useful for taking huge CSV files that do not fit into RAM and store them as a SQLite db which can be streamed via dplyr. This way you can work with really big data files without being limited by your machines RAM.

csvToSQLite(
  csv_file,
  sqlite_file,
  table_name,
  pre_process_size = 1000,
  chunk_size = 50000,
  delim = ","
)

Arguments

csv_file
name of the CSV file to convert

sqlite_file
name of the newly created sqlite file

table_name
name of the table to store the data table in the sqlite dbase

pre_process_size
the number of lines to check the data types of the individual columns (default 1000)

chunk_size
the number of lines to read for each chunk (default 50000)

delim
the field delimiter to use (default ,)

Examples

library(RSQLite)
library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

sqlite_file <- "example.sqlite"
table_name <- "example"
write.csv(airquality, "example.csv", row.names=FALSE)
csvToSQLite("example.csv", sqlite_file, table_name,
              pre_process_size = 1000, chunk_size = 50000)

#> Parsed with column specification:
#> cols(
#>   Ozone = col_double(),
#>   Solar.R = col_double(),
#>   Wind = col_double(),
#>   Temp = col_double(),
#>   Month = col_double(),
#>   Day = col_double()
#> )

mydb <- src_sqlite(sqlite_file, create = FALSE)

#> Warning: `src_sqlite()` is deprecated as of dplyr 1.0.0.
#> Please use `tbl()` directly with a database connection
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_warnings()` to see where this warning was generated.

mydata <- tbl(mydb, table_name)
head(mydata)

#> # Source:   lazy query [?? x 6]
#> # Database: sqlite 3.30.1
#> #   [/home/michael/src/datools/docs/reference/example.sqlite]
#>   Ozone Solar.R  Wind  Temp Month   Day
#>   <dbl>   <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1    41     190   7.4    67     5     1
#> 2    36     118   8      72     5     2
#> 3    12     149  12.6    74     5     3
#> 4    18     313  11.5    62     5     4
#> 5    NA      NA  14.3    56     5     5
#> 6    28      NA  14.9    66     5     6