Lag variables in data nlags

This function uses an existing dataset in the shape of a tibble and then lags each variable(column) in the tibble from 1 all the way up to nlag.

lagvariables(x, nlag = 5)

Arguments

x
the data tibble to operate on

nlag
the max number of lags to create for each variable

Value

a new tibble with the data lagged

Details

This function requires that the first column in your tibble is called "date" and represents some kind of time epoch. Note also that you will get no NA's returned. This means that the returned data will start nlag-1 days later than the input data.

Examples

library(dplyr)
tmpdf <- tibble(
  date = seq(from = as.Date("2018-01-01"), length.out = 10, by = "1 day"),
  a = 1:10, b = 11:20, d = 21:30
)
lagdf <- tmpdf %>% lagvariables(4)
lagdf

#> # A tibble: 7 x 13
#>   date    a_l0  a_l1  a_l2  a_l3  b_l0  b_l1  b_l2  b_l3  d_l0  d_l1  d_l2  d_l3
#>   <chr>  <int> <int> <int> <int> <int> <int> <int> <int> <int> <int> <int> <int>
#> 1 2018-…     4     3     2     1    14    13    12    11    24    23    22    21
#> 2 2018-…     5     4     3     2    15    14    13    12    25    24    23    22
#> 3 2018-…     6     5     4     3    16    15    14    13    26    25    24    23
#> 4 2018-…     7     6     5     4    17    16    15    14    27    26    25    24
#> 5 2018-…     8     7     6     5    18    17    16    15    28    27    26    25
#> 6 2018-…     9     8     7     6    19    18    17    16    29    28    27    26
#> 7 2018-…    10     9     8     7    20    19    18    17    30    29    28    27

tmpdf

#> # A tibble: 10 x 4
#>    date           a     b     d
#>    <date>     <int> <int> <int>
#>  1 2018-01-01     1    11    21
#>  2 2018-01-02     2    12    22
#>  3 2018-01-03     3    13    23
#>  4 2018-01-04     4    14    24
#>  5 2018-01-05     5    15    25
#>  6 2018-01-06     6    16    26
#>  7 2018-01-07     7    17    27
#>  8 2018-01-08     8    18    28
#>  9 2018-01-09     9    19    29
#> 10 2018-01-10    10    20    30