Make waterfall data tibble from a tibble of cash flows

The only requirement for this function to work is that the first row and last row of the tibble represent the cash in and cash out. In other words the Base and the Total when talking about sales modeling.

to_waterfall(data)

Arguments

data
the data tibble to operate on

Value

a new data tibble with columns necessary to create a waterfall visualization

Examples

datadf <-
  tibble::tibble(
    desc = c(
      "starting_net_income",
      "write_offs",
      "sales",
      "COGS",
      "overhead",
      "taxes",
      "bonuses",
      "other_income",
      "ending_net_income"
    ),
    amount = c(1200, -2300, 11000, -9000, -1900, -1200, -3000, 7000, 1800)
  )
to_waterfall(datadf)

#> # A tibble: 9 x 6
#>      id desc                type  start   end amount
#>   <int> <fct>               <fct> <dbl> <dbl>  <dbl>
#> 1     1 starting_net_income net       0  1200   1200
#> 2     2 write_offs          out    1200 -1100  -2300
#> 3     3 sales               in    -1100  9900  11000
#> 4     4 COGS                out    9900   900  -9000
#> 5     5 overhead            out     900 -1000  -1900
#> 6     6 taxes               out   -1000 -2200  -1200
#> 7     7 bonuses             out   -2200 -5200  -3000
#> 8     8 other_income        in    -5200  1800   7000
#> 9     9 ending_net_income   net    1800     0   1800

# Change order
datadf2 <- to_waterfall(datadf)
firstindex <- which(datadf$desc == "sales")
datadf2[c(firstindex, setdiff(seq_along(datadf$amount), firstindex)), ]

#> # A tibble: 9 x 6
#>      id desc                type  start   end amount
#>   <int> <fct>               <fct> <dbl> <dbl>  <dbl>
#> 1     3 sales               in    -1100  9900  11000
#> 2     1 starting_net_income net       0  1200   1200
#> 3     2 write_offs          out    1200 -1100  -2300
#> 4     4 COGS                out    9900   900  -9000
#> 5     5 overhead            out     900 -1000  -1900
#> 6     6 taxes               out   -1000 -2200  -1200
#> 7     7 bonuses             out   -2200 -5200  -3000
#> 8     8 other_income        in    -5200  1800   7000
#> 9     9 ending_net_income   net    1800     0   1800